+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 #### Fixed
 - fix usage of `eval(repr(Editop))` and `eval(repr(Opcode))`
+- fix opcode conversion for empty source sequence
 
 ### [2.2.0] - 2022-07-19
 #### Changed